requirements.txt
python-dotenv
fastapi
uvicorn
langchain-openai
langchain-chroma
langchain-community
pypdf
pandas
python-pptx
python-docx
python-multipart